feature: arquitetura banco dados

+++ src/main/java/com/cit/virtual_ponto/cadastro_funcionarios/repositories/EmpresaRepository.java
@@ -1,7 +1,8 @@
 package com.cit.virtual_ponto.cadastro_funcionarios.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.cit.virtual_ponto.cadastro_funcionarios.models.PessoaJuridica;
 
-public interface EmpresaRepository extends JpaRepository<PessoaJuridica, Long> {
+import com.cit.virtual_ponto.cadastro_funcionarios.models.pessoa.PessoaJuridica;
+
+public interface EmpresaRepository extends JpaRepository<PessoaJuridica, Integer> {
 }
